#28 [REFACTOR] 세미나 조회 페이지 컴포넌트 분리
